+component -> admin delete ^erd ^sql_init *refactoring

+++ public/scripts/deleteFromComponentPage.js
@@ -1,10 +1,13 @@
 import {deleteComponent} from './deleteComponent.js';
 
 document.addEventListener('DOMContentLoaded', () => {
-    document.querySelector('.delete').addEventListener('click', async (event) => {
-            const componentId = document.querySelector('.delete').getAttribute('data-component-id');
-            await deleteComponent(componentId, null);
-            window.location.replace('/collection');
-        }
-    );
+    const deleteButton = document.querySelector('.delete');
+    if (deleteButton) {
+        deleteButton.addEventListener('click', async (event) => {
+                const componentId = deleteButton.getAttribute('data-component-id');
+                await deleteComponent(componentId, null);
+                window.location.replace('/collection');
+            }
+        );
+    }
 });
